+++ crates/celox-frontend-veryl/Cargo.toml
@@ -13,6 +13,8 @@ celox-sir      = { path = "../celox-sir" }
 celox-slt      = { path = "../celox-slt" }
 fxhash         = { workspace = true }
 miette         = { workspace = true }
+num-bigint     = { workspace = true }
+num-traits     = { workspace = true }
 thiserror      = { workspace = true }
 veryl-analyzer = { workspace = true }
 veryl-metadata = { workspace = true }
